requirements.txt
asgiref==3.3.4
click==8.0.1
Django==3.2.3
django-cleanup==5.2.0
django-cors-headers==3.7.0
djangorestframework==3.12.4
dlib==19.22.0
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.20.3
opencv-python==4.5.2.52
Pillow==8.2.0
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
